FoReco: point forecast reconciliation

An R package offering classical (bottom-up and top-down), and modern (optimal and heuristic combination) forecast reconciliation procedures for cross-sectional, temporal, and cross-temporal linearly constrained time series.

Details

The FoReco package is designed for point forecast reconciliation, a post-forecasting process aimed to improve the accuracy of the base forecasts for a system of linearly constrained (e.g. hierarchical/grouped) time series. The main functions are:

htsrec():

cross-sectional (contemporaneous) forecast reconciliation.

thfrec():

forecast reconciliation for a single time series through temporal hierarchies.

lccrec():

level conditional forecast reconciliation for genuine hierarchical/grouped time series.

tdrec():

top-down (cross-sectional, temporal, cross-temporal) forecast reconciliation for genuine hierarchical/grouped time series.

ctbu():

bottom-up cross-temporal forecast reconciliation.

tcsrec():

heuristic first-temporal-then-cross-sectional cross-temporal forecast reconciliation.

cstrec():

heuristic first-cross-sectional-then-temporal cross-temporal forecast reconciliation.

iterec():

heuristic iterative cross-temporal forecast reconciliation.

octrec():

optimal combination cross-temporal forecast reconciliation.

References

Di Fonzo, T., and Girolimetto, D. (2021), Cross-temporal forecast reconciliation: Optimal combination method and heuristic alternatives, International Journal of Forecasting, in press doi:10.1016/j.ijforecast.2021.08.004 .

Di Fonzo, T., Girolimetto, D. (2022), Forecast combination based forecast reconciliation: insights and extensions, International Journal of Forecasting, in press.

Author

Tommaso Di Fonzo and Daniele Girolimetto, Department of Statistical Sciences, University of Padua (Italy).