Cross-temporal reconciliation tools

Some useful tools for the cross-temporal forecast reconciliation of a linearly constrained (hierarchical/grouped) multiple time series.

ctf_tools(C, m, h = 1, Ut, nb, sparse = TRUE)

Arguments

C

(\(n_a \times n_b\)) cross-sectional (contemporaneous) matrix mapping the bottom level series into the higher level ones.

m

Highest available sampling frequency per seasonal cycle (max. order of temporal aggregation, \(m\)), or a subset of the \(p\) factors of \(m\).

h

Forecast horizon for the lowest frequency (most temporally aggregated) time series (default is 1).

Ut

Zero constraints cross-sectional (contemporaneous) kernel matrix \((\mathbf{U}'\mathbf{y} = \mathbf{0})\) spanning the null space valid for the reconciled forecasts. It can be used instead of parameter C, but nb (\(n = n_a + n_b\)) is needed if \(\mathbf{U}' \neq [\mathbf{I} \ -\mathbf{C}]\). If the hierarchy admits a structural representation, \(\mathbf{U}'\) has dimension (\(n_a \times n\)).

nb

Number of bottom time series; if C is present, nb and Ut are not used.

sparse

Option to return sparse object (default is TRUE).

Value

ctf list with:

Ht

Full row-rank cross-temporal zero constraints (kernel) matrix coherent with \(\mathbf{y} = \mbox{vec}(\mathbf{Y}')\): \(\mathbf{H}'\mathbf{y} = \mathbf{0}\).

Hbrevet

Complete, not full row-rank cross-temporal zero constraints (kernel) matrix coherent with \(\mathbf{y} = \mbox{vec}(\mathbf{Y}')\): \(\breve{\mathbf{H}}'\mathbf{y} = \mathbf{0}\).

Hcheckt

Full row-rank cross-temporal zero constraints (kernel) matrix coherent with \(\check{\mathbf{y}}\) (structural representation): \(\check{\mathbf{H}}' \check{\mathbf{y}} = \mathbf{0}\).

Ccheck

Cross-temporal aggregation matrix \(\check{\mathbf{C}}\) coherent with \(\check{\mathbf{y}}\) (structural representation).

Scheck

Cross-temporal summing matrix \(\check{\mathbf{S}}\) coherent with \(\check{\mathbf{y}}\) (structural representation).

Fmat

Cross-temporal summing matrix \(\widetilde{\mathbf{F}}\) coherent with \(\mathbf{y} = \mbox{vec}(\mathbf{Y}')\).

hts list from hts_tools .

thf list from thf_tools .

See also

Examples

# One level hierarchy (na = 1, nb = 2) with quarterly data
obj <- ctf_tools(C = matrix(c(1, 1), 1), m = 4)