Level conditional forecast reconciliation for genuine hierarchical/grouped time series

Forecast reconciliation procedure based on and extending the original proposal by Hollyman et al. (2021). Unbiased top-down and level conditional reconciled forecasts may be computed in cross-sectional, temporal, and cross-temporal frameworks. The reconciled forecasts are conditional to (i.e., constrained by) the base forecasts of the upper levels of the hierarchy (exogenous constraints). The linear constraints linking the variables may be dealt with endogenously as well, according to Di Fonzo (2021). Combined Conditional Coherent (CCC, Hollyman et al., 2021) forecasts may be calculated as simple averages of different middle-out level reconciled forecasts, with both endogenous and exogenous constraints.

levrec(
  basef,
  C,
  nk,
  m,
  weights,
  bmeans = NULL,
  const = "exogenous",
  CCC = TRUE,
  nn = FALSE,
  settings = list(),
  ...
)

Arguments

basef
(\(h \times n\)) matrix of base forecasts to be reconciled; h is the forecast horizon and n is the total number of time series.

C
(\(n_a \times n_b\)) cross-sectional (contemporaneous) matrix mapping the bottom level series into the higher level ones (or a list of matrices forming \(\textbf{C} = [\textbf{C}_1' \; \textbf{C}_2' \; ... \; \textbf{C}_K']'\), \(1, ..., K\) being the number of upper levels, only cross-sectional).

nk
(\(K \times 1\)) vector containing the number of time series in each level of the hierarchy (nk[1] = 1).

m
Highest available sampling frequency per seasonal cycle (max. order of temporal aggregation, \(m\)), or a subset of \(p\) factors of \(m\).

weights
covariance matrix or a vector

bmeans
(\(h \times n_b\)) matrix of bts means (prova)

const
exogenous or endogenous constraints (default is "exo")

CCC
Option to return Combined Conditional Coherent reconciled forecasts (default is TRUE).

nn
Logical value: TRUE if non-negative reconciled forecasts are wished.

settings
Settings for osqp (object osqpSettings). The default options are: verbose = FALSE, eps_abs = 1e-5, eps_rel = 1e-5, polish_refine_iter = 100 and polish = TRUE. For details, see the osqp documentation (Stellato et al., 2019).

...
additional functional arguments passed to htsrec of FoReco (type, nn, nn_type, bounds)

Value

The function returns the unbiased top-down reconciled forecasts in case of an elementary hierarchy with one level. Otherwise it returns a list with

recf

the Unbiased Top-Down reconciled forecasts, if CCC = FALSE, or the CCC reconciled forecasts.

lev

list of level conditional reconciled forecasts (Unbiased Top-Down + Middle-Out + BU).

info (type="osqp")

matrix with some useful indicators (columns) for each forecast horizon \(h\) (rows): run time (run_time), number of iteration, norm of primal residual (pri_res), status of osqp's solution (status) and polish's status (status_polish).

References

Di Fonzo, T., Girolimetto, D. (2020), Cross-Temporal Forecast Reconciliation: Optimal Combination Method and Heuristic Alternatives, Department of Statistical Sciences, University of Padua, arXiv:2006.08570.

Di Fonzo, T. (2021), Approfondimenti, interpretazione ed estensioni della procedura Unbiased Top-Down Forecast Reconciliation di Hollyman et al. (2021) (in Italian, mimeo).

Hollyman, R., Petropoulos, F., Tipping, M.E. (2021), Understanding Forecast Reconciliation, European Journal of Operational Research (in press).

Examples

data(FoReco_data)
# Cross sectional aggregation matrix
C <- FoReco_data$C
# monthly base forecasts
id <- which(simplify2array(strsplit(colnames(FoReco_data$base), split = "_"))[1, ] == "k1")
mbase <- t(FoReco_data$base[, id])
# residuals
id <- which(simplify2array(strsplit(colnames(FoReco_data$res), split = "_"))[1, ] == "k1")
mres <- t(FoReco_data$res[, id])
# covariance matrix of the bts base forecasts, computed using the in-sample residuals
Wres <- cov(mres)
# diagonal matrix containing the residual variances of the bts base forecasts
Wb <- Wres[c("AA","AB", "BA", "BB", "C"),
           c("AA","AB", "BA", "BB", "C")]

## EXOGENOUS CONSTRAINTS AND DIAGONAL COVARIANCE MATRIX (Hollyman et al., 2021)
# Unbiased Top-Down reconciliation for an elementary hierarchy:
# 1 top-level series + nb bottom-level series (Level 2 base forecasts are not considered).
# The input is given by the base forecasts of the top and bottom levels series,
# along with a vector of positive weights for the bts base forecasts
exo_EHd <- levrec(basef = mbase[, c("T","AA","AB", "BA", "BB", "C")], weights = Wb)

# Level conditional reconciled forecasts
# L1: Unbiased Top-Down reconciled forecasts for the whole hierarchy
# L2: Middle-Out reconciled forecasts using Level 2 conditional reconciled forecasts
# L3: Bottom-Up reconciled forecasts
exo_LCd <- levrec(basef = mbase, C = C, nk = c(1,2), weights = diag(Wb), CCC = FALSE)

# Combined Conditional Coherent (CCC) reconciled forecasts
# recf: CCC reconciled forecasts matrix
# L1: Unbiased Top-Down reconciled forecasts for the whole hierarchy
# L2: Middle-Out reconciled forecasts using Level 2 conditional reconciled forecasts
# L3: Bottom-Up reconciled forecasts
exo_CCCd <- levrec(basef = mbase, C = C, nk = c(1,2), weights = diag(Wb))

## EXOGENOUS CONSTRAINTS AND FULL COVARIANCE MATRIX
# Simply substitute weights=diag(Wb) with weights=Wb
# Warning: possible negative reconciled forecasts may appear.
# This issue must be fixed in the next future.
exo_EHf <- levrec(basef = mbase[, c("T","AA","AB", "BA", "BB", "C")], weights = Wb)
exo_LCf <- levrec(basef = mbase, C = C, nk = c(1,2), weights = Wb, CCC = FALSE)
exo_CCCf <- levrec(basef = mbase, C = C, nk = c(1,2), weights = Wb)

## ENDOGENOUS CONSTRAINTS AND DIAGONAL COVARIANCE MATRIX
# parameters of function htsrec(), like type, nn, nn_type, bounds, may be used

# Elementary hierarchy:
endo_EHd <- levrec(basef = mbase[, c("T","AA","AB", "BA", "BB", "C")],
                  weights = diag(Wres)[c("T","AA","AB", "BA", "BB", "C")],
                  const = "endo")

# Level conditional endogenously reconciled forecasts
# L1: Middle-Out reconciled forecasts for the whole hierarchy using Level 1
# L2: Middle-Out reconciled forecasts for the whole hierarchy using Level 2
# L3: Bottom-Up reconciled forecasts
endo_LCd <- levrec(basef = mbase, C = C, nk = c(1,2), weights = diag(Wres),
                  const = "endo", CCC = FALSE)

# Endogenously Combined Conditional Coherent (CCC) reconciled forecasts
# recf: endogenously CCC reconciled forecasts matrix
# lev: list of reconciled forecasts' matrices
# L1: Middle-Out reconciled forecasts for the whole hierarchy using Level 1
# L2: Middle-Out reconciled forecasts for the whole hierarchy using Level 2
# L3: Bottom-Up reconciled forecasts
endo_CCCd <- levrec(basef = mbase, C = C, nk = c(1,2),
                    weights = diag(Wres), const = "endo")

## ENDOGENOUS CONSTRAINTS AND FULL COVARIANCE MATRIX
# Simply substitute weights=diag(Wres) with weights=Wres
endo_EHf <- levrec(basef = mbase[, c("T","AA","AB", "BA", "BB", "C")],
                   weights = Wres[c("T","AA","AB", "BA", "BB", "C")],
                   const = "endo")
endo_LCf <- levrec(basef = mbase, C = C, nk = c(1,2),
                   weights = Wres, const = "endo", CCC = FALSE)
endo_CCCf <- levrec(basef = mbase-40, C = C, nk = c(1,2),
                   weights = Wres, const = "endo")