Informations on the reconciliation process

This function extracts reconciliation information from the output of any reconciled function implemented by FoReco.

Usage

recoinfo(x, verbose = TRUE)

Arguments

x

An output from any reconciliation function implemented by FoReco.

verbose

If TRUE (defaults), reconciliation information are printed.

Value

A list containing the following reconciliation process informations:

rfun

the reconciliation function.

cs_n

the cross-sectional number of variables.

te_set

the set of temporal aggregation orders.

forecast_horizon

the forecast horizon (in temporal and cross-temporal frameworks, for the most temporally aggregated series).

framework

the reconciliation framework (cross-sectional, temporal or cross-temporal).

info

non-negative reconciled forecast convergence information.

lcc

list of level conditional reconciled forecasts (+ BU) for cslcc, telcc and ctlcc.

nn

if TRUE, all the forecasts are not negative.

comb

the covariance approximation.

See also