FoReco: point forecast reconciliation

An R package offering classical (bottom-up), optimal and heuristic combination forecast reconciliation procedures for cross-sectional, temporal, and cross-temporal linearly constrained time series.

Details

The FoReco package is designed for point forecast reconciliation, a post-forecasting process aimed to improve the quality of the base forecasts for a system of linearly constrained (e.g. hierarchical/grouped) time series. The main functions are:

htsrec():

cross-sectional (contemporaneous) forecast reconciliation.

thfrec():

forecast reconciliation for a single time series through temporal hierarchies.

tcsrec():

heuristic first-temporal-then-cross-sectional cross-temporal forecast reconciliation.

cstrec():

heuristic first-cross-sectional-then-temporal cross-temporal forecast reconciliation.

iterec():

heuristic iterative cross-temporal forecast reconciliation.

octrec():

optimal combination cross-temporal forecast reconciliation.

References

Di Fonzo, T., Girolimetto, D. (2020), Cross-Temporal Forecast Reconciliation: Optimal Combination Method and Heuristic Alternatives, Department of Statistical Sciences, University of Padua, arXiv:2006.08570.