Shrinkage of the covariance matrix

Shrinkage of the covariance matrix according to Schäfer and Strimmer (2005).

shrink_estim(x, minT = T)

Arguments

x

residual matrix

minT

this param allows to calculate the covariance matrix according to the original hts formulation (TRUE) or according to the standard approach (FALSE).

Value

A list with two objects: the first ($scov) is the shrunk covariance matrix and the second ($lambda) is the shrinkage intensity coefficient.

References

Schäfer, J.L., Strimmer, K. (2005), A Shrinkage Approach to Large-Scale Covariance Matrix Estimation and Implications for Functional Genomics, Statistical Applications in Genetics and Molecular Biology, 4, 1

Hyndman, R. J., Lee, A., Wang, E., and Wickramasuriya, S. (2020). hts: Hierarchical and Grouped Time Series, R package version 6.0.1, https://CRAN.R-project.org/package=hts.

See also

Author

This function is a modified version of the shrink_estim() hidden function of hts.