Cross-sectional 'structural representation' of a general linearly constrained multiple time series

Switching from a zero-constraints kernel representation of a linearly constrained multiple time series, \(\mathbf{U}'\mathbf{y}=\mathbf{0}\), to a 'structural representation', \(\mathbf{y}=\mathbf{S}\mathbf{b}\). (Experimental version)

ut2c(Ut, sparse = TRUE, verbose = FALSE)

Arguments

Ut

(\(r \times c\)) zero constraints cross-sectional (contemporaneous) kernel matrix \((\mathbf{U}'\mathbf{y} = \mathbf{0})\) spanning the null space valid for the target forecasts.

sparse

Option to return sparse object (default is TRUE).

verbose

If TRUE, print intermediate steps of srref.

Value

A list with

Ut_reshape

matrix with rows and columns rearranged so that each row has 1 as the first non-null element starting from the left, and that the first r columns have 1 as the first element starting from the bottom, \(\mathbf{U}_{reshape}' = \mathbf{U}'[,\mbox{cid}]\).

Ut_rref

reduced row echelon form of \(\mathbf{U}'_{reshape}\) with srref.

C

(\(n_a \times n_b\)) cross-sectional (contemporaneous) matrix mapping the bottom level series into the higher level ones, \(\mathbf{U}_{srref}' = [\mathbf{I} \ -\mathbf{C}]\).

cid

(\(c \times 1\)) vector of the column permutations of the matrix \(\mathbf{U}'\).

nb

number of bottom time series, \(n_b\).

na

number of upper time series, \(n_a = r\).

n

number of time series, \(n_a + n_b = n = c\).

Details

Consider the simple example of linearly constrained multiple time series consisting of two hierarchies, each with distinct bottom time series, with a common top-level series (\(T\)): \[\begin{array}{ll} 1)\; T = A + B & 4)\; T = X + Y \cr 2)\; A = AA + AB & 5)\; X = XX + XY \cr 3)\; B = BA + BB + BC & 6)\; Y = YX + YY \end{array}.\] Given the cross-sectional aggregation matrices of each hierarchy, \[\mathbf{C}_1 = \left[\begin{array}{ccccc} 1 & 1 & 0 & 0 & 0\cr 0 & 0 & 1 & 1 & 1 \end{array}\right]\quad \mathrm{and} \quad \mathbf{C}_2 = \left[\begin{array}{ccccc} 1 & 1 & 0 & 0\cr 0 & 0 & 1 & 1 \end{array}\right],\] the zero constraints cross-sectional kernel matrix \(\mathbf{U}'\), which accounts for the constraints of both hierarchies, can be built as follows: \[\mathbf{U}' = \left[\begin{array}{cccccccccccccc|c} 1 & 0 & 0 &-1 &-1 &-1 &-1 &-1 & 0 & 0 & 0 & 0 & 0 & 0 & T\cr 1 & 0 & 0 & 0 & 0 & 0 & 0 & 0 & 0 & 0 &-1 &-1 &-1 &-1 & T\cr 0 & 1 & 0 &-1 &-1 & 0 & 0 & 0 & 0 & 0 & 0 & 0 & 0 & 0 & A\cr 0 & 0 & 1 & 0 & 0 &-1 &-1 &-1 & 0 & 0 & 0 & 0 & 0 & 0 & B\cr 0 & 0 & 0 & 0 & 0 & 0 & 0 & 0 & 1 & 0 &-1 &-1 & 0 & 0 & X\cr 0 & 0 & 0 & 0 & 0 & 0 & 0 & 0 & 0 & 1 & 0 & 0 &-1 &-1 & Y\cr \hline T & A & B & AA& AB& BA& BB& BC& X & Y & XX& XY& YX& YY& \end{array}\right].\] Function ut2c returns a matrix \(\mathbf{U}'_{rref} = \left[\mathbf{I}_6 \; -\mathbf{C} \right]\): \[\mathbf{U}'_{rref} = \left[\begin{array}{cccccccccccccc|c} 1 & 0 & 0 & 0 & 0 & 0 & 0 & 0 & 0 & 0 &-1 &-1 &-1 &-1 & T\cr 0 & 1 & 0 & 0 & 0 & 0 & 0 & 1 & 1 & 1 &-1 &-1 &-1 &-1 & A\cr 0 & 0 & 1 & 0 & 0 & 0 & 0 &-1 &-1 &-1 & 0 & 0 & 0 & 0 & B\cr 0 & 0 & 0 & 1 & 0 & 0 & 0 & 0 & 0 & 0 &-1 &-1 & 0 & 0 & X\cr 0 & 0 & 0 & 0 & 1 & 0 & 0 & 0 & 0 & 0 & 0 & 0 &-1 &-1 & Y\cr 0 & 0 & 0 & 0 & 0 & 1 & 1 & 1 & 1 & 1 &-1 &-1 &-1 &-1 & AA\cr \hline T & A & B & X & Y & AA& AB& BA& BB& BC& XX& XY& YX& YY& \end{array}\right],\] from which the 'structural sum matrix' \[\mathbf{S} = \left[\mathbf{C}' \; \mathbf{I}_8 \right]'\] may be easily obtained.

References

Di Fonzo, T., Girolimetto, D. (2020), Cross-Temporal Forecast Reconciliation: Optimal Combination Method and Heuristic Alternatives, Department of Statistical Sciences, University of Padua, arXiv:2006.08570.

See also

Examples

C1 <- Matrix(c(1,1,0,0,0,
               0,0,1,1,1), 2, byrow = TRUE, sparse = TRUE)
C2 <- Matrix(c(1,1,0,0,
               0,0,1,1), 2, byrow = TRUE, sparse = TRUE)
Ut <- rbind(c(1, rep(0, NROW(C1)), rep(-1, NCOL(C1)), rep(0, NROW(C2)), rep(0, NCOL(C2))),
            c(1, rep(0, NROW(C1)), rep(0, NCOL(C1)), rep(0, NROW(C2)), rep(-1, NCOL(C2))),
            cbind(0, Diagonal(NROW(C1)), -C1, Matrix(0, NROW(C1), NROW(C2)+NCOL(C2))),
            cbind(0, Matrix(0, NROW(C2), NROW(C1)+NCOL(C1)), Diagonal(NROW(C2)), -C2))
colnames(Ut) <- c("T", "A", "B", "AA", "AB", "BA", "BB", "BC",
                  "X", "Y", "XX", "XY", "YX", "YY")
rownames(Ut) <- c("T", "T", "A", "B", "X", "Y")
obj_ut2c <- ut2c(Ut, sparse = FALSE)
#> Error in ut2c(Ut, sparse = FALSE): could not find function "ut2c"
Ut_struc <- obj_ut2c$Ut_rref
#> Error in eval(expr, envir, enclos): object 'obj_ut2c' not found
S <- rbind(obj_ut2c$C, diag(1, obj_ut2c$nb))
#> Error in rbind(obj_ut2c$C, diag(1, obj_ut2c$nb)): object 'obj_ut2c' not found